Test Framework changes to incorporate Simba ODBC Driver into test automation framework.
Added odbcTest.py program to execute query using ODBC driver and generate output file.
The program also generates a file containing the names of the columns in the output,
and the data types of the columns in the output.  These three files (output, column names,
data types) are used to verify that the output from the ODBC driver matches the baseline
output in the Test Framework (generated using Apache JDBC driver), when possible.
Exceptions are handled in other parts of the Test Framework using the same technique
used for Simba JDBC driver results.

Updated DrillTestOdbc to use the odbcTest.py program, and to read in the contents of the
three files generated by odbcTest.py (output, column  names, data types).  Also updated
DrillTestOdbc to reflect some enhancements made to DrillTestJdbc.