Added a new failure type, ORDERBY_FAILURE, which occurs if a SQL statement has an order-by clause that cannot be validated by the Test Framework.  Please see the README.md file for information on how to construct a SQL statement with an order-by clause that can be validated by the Test Framework.
Added a new verification method, "text", which checks if the actual output
and expected output are exactly the same, byte by byte.  It verifies content
as well as order.  Every row that is returned must be returned in the same
order as in the expected output.

Added a whitelist file for orderby tests.  Tests that are listed in this file
are SQL statements with an order-by clause.  The order-by clause cannot be
validated.  Instead of marking these tests as failed, they are PASSED until
they can be fixed in the future.  There are 302 tests that are currently
failing at the time of this commit.  They will need to be fixed at some
piont in the future.