Separate data verification failures from plan verification failures. (#315)
Add query ID to the status of each test.  The query ID can be used to find
the profile.  Each node has to be checked for the profile.  Here is a sample:

   PASS (1.988 s) /root/drill-test-framework/framework/resources/Functional/window_functions/lag_func/lag_Fn_93.q (connection: 2007519911) (queryID: 26504575-863a-85bc-85d4-ae7a0ef3a8b9)

Execution failures and Timeouts may not have a query ID because the query did not complete.

With this check-in, the test framework needs to be compiled with the Apache JDBC driver.
It can then be used with the Apache or Simba JDBC driver.