Add option to enable test framework to run tests in a specified order. (#328)
The option is enabled with the -rR flag, for repeatRun.
If specified for the first time, it will create a file in framework/resources/queryLog.log
that contains the tests that are run, in the order they are run.
Once this file is created, the next run will run the tests in the order specified
in the file.  The file can also be created manually.
This implementation requires that the tests to be run are also specified by the -s option.
It takes the tests that are normally collected, and re-orders the tests in the order
specified in the queryLog.log file.
This is useful for failures that occur when tests are run in a particular order.
It is mainly useful if the tests are run without concurrency, meaning one at a time.