1.12.0 commits cherrypick (#392)
* MD-3291: Removed usage of paths outside current schema (#371)

* DRILL-5795: Make test more robust for data addition in the plan (#352)

DRILL-5795 updates the plan to add the number of rowgroup scanned
because the number of file scanned is less relevant. q28.e was
expecting the number of files and the useMetadata to be next to each
other. This is not the case anymore as the number of rowgroup is
in between. For robustness of the test, it is better not to assume they
are next to each other and in other tests, it is not expected.

This PR changes the expected string to a more robust one, so that it
works before and after the integration of DRILL-5795.

* MD-2916: Fixed baseline failures caused by column reordering (#351)

* MD-2296: Changes to schema change due to empty batch test cases (#327)

* MD-2916: Fixed baseline failure caused by column reordering