Cherry-pick framework changes (#422)
* DRILL-4994: use prepareStatement to test pre-check query (#359)

* Improve support for comments in tests (#369)

* Add option to enable test framework to run tests in a specified order. (#328)

The option is enabled with the -rR flag, for repeatRun.
If specified for the first time, it will create a file in framework/resources/queryLog.log
that contains the tests that are run, in the order they are run.
Once this file is created, the next run will run the tests in the order specified
in the file.  The file can also be created manually.
This implementation requires that the tests to be run are also specified by the -s option.
It takes the tests that are normally collected, and re-orders the tests in the order
specified in the queryLog.log file.
This is useful for failures that occur when tests are run in a particular order.
It is mainly useful if the tests are run without concurrency, meaning one at a time.

* Test Framework changes to incorporate Simba JDBC Driver into test automation framework. (#254)

Added "-b" option to indicate file extension for other drivers such as Simba JDBC.
Re-arranged command line options in CmdParam in alphabetical order.
Added code to ColumnList to convert Simba JSON records which are multi-line records
into single-line records.
Added support for SQL REAL data type.  Simba uses REAL instead of FLOAT.
Added Simba class to ConnectionPool for refereence.
Changed import for String in ShellRunner to use original Google String class.
Added expected file name to Results to help with analyzing results for failed tests.
Added code to getExpectedFile to look for either original expected file,
expected file with Simba JDBC results, or failed expected file indicating that the
query file cannot be run with Simba JDBC.
Modified code that runs failed tests to include query file extension.

* Test Framework changes to incorporate Simba ODBC Driver into test automation framework. (#386)

Added odbcTest.py program to execute query using ODBC driver and generate output file.
The program also generates a file containing the names of the columns in the output,
and the data types of the columns in the output.  These three files (output, column names,
data types) are used to verify that the output from the ODBC driver matches the baseline
output in the Test Framework (generated using Apache JDBC driver), when possible.
Exceptions are handled in other parts of the Test Framework using the same technique
used for Simba JDBC driver results.

Updated DrillTestOdbc to use the odbcTest.py program, and to read in the contents of the
three files generated by odbcTest.py (output, column  names, data types).  Also updated
DrillTestOdbc to reflect some enhancements made to DrillTestJdbc.

* Console Output Clean-up (Part 1) (#418)